+++ backend/src/auth/controller/auth.controller.ts
@@ -6,8 +6,6 @@ import {
   Req,
   Res,
   UseGuards,
-  Post,
-  Body,
   Query
 } from '@nestjs/common';
 import { AuthService } from '../service/auth.service';
@@ -106,23 +104,5 @@ export class AuthController {
   @UseGuards(AuthGuard('apple'))
   async appleOAuthCallback(@Req() req) {
     return await this.authService.handleAppleOAuthLogin(req);
-  }
-
-  // New API endpoint for social account linking
-  @Post('/link-social')
-  @UseGuards(AuthGuard('jwt')) // Assuming JWT authentication for authorization
-  async linkSocialAccount(
-    @Body() { userId, provider, accessToken }: { userId: string; provider: string; accessToken: string },
-  ) {
-    try {
-      const linkedUser = await this.authService.linkSocialAccount(userId, provider, accessToken);
-      return linkedUser; // Return success response or relevant data
-    } catch (error) {
-      if (error instanceof HttpException) {
-        throw error; // Re-throw existing HttpExceptions for consistent error handling
-      } else {
-        throw new HttpException('Failed to link social account', HttpStatus.INTERNAL_SERVER_ERROR);
-      }
-    }
   }
 }
