+++ backend/src/auth/service/auth.service.ts
@@ -122,33 +122,4 @@ export class AuthService {
       },
     });
   }
-
-  async linkSocialAccount(userId: string, provider: string, accessToken: string) {
-    const user = await this.findUserById(userId);
-    if (!user) {
-      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
-    }
-
-    let email: string | undefined;
-
-    const existingUser = await this.findUserByEmail(email);
-    if (existingUser && existingUser.id !== user.id) {
-      throw new HttpException('Social account already linked to a different user', HttpStatus.CONFLICT);
-    }
-
-    user.socialAccounts = user.socialAccounts || [];
-    user.socialAccounts.push({ provider, accessToken });
-
-    await this.updateUser(user);
-
-    return user;
-  }
-
-  private async findUserById(id: string) {
-    return await this.prisma.user.findUnique({ where: { id } });
-  }
-
-  private async updateUser(user: any) {
-    return await this.prisma.user.update({ where: { id: user.id }, data: user });
-  }
 }
