+++ backend/src/user/controller/user.controller.ts
@@ -1,6 +1,8 @@
-import { Body, Controller, Get, Param, Post } from '@nestjs/common';
+import { Body, Controller, Get, Param, Post,
+  UseGuards, HttpException, HttpStatus } from '@nestjs/common';
 import { CurrentUser } from '../../decorators/current-user.decorator';
 import { UserService } from '../service/user.service';
+import { AuthGuard } from '@nestjs/passport';
 import { User } from '@prisma/client';
 import { RatingDto } from '../dto/rating.dto';
 
@@ -33,5 +35,22 @@ export class UserController {
     @Param('revieweeUserId') revieweeUserId: string,
   ) {
     return this.userService.getUserReviews(user, false, revieweeUserId);
+  }
+
+  @Post('/link-social')
+  @UseGuards(AuthGuard('jwt')) 
+  async linkSocialAccount(
+    @Body() { userId, provider, accessToken }: { userId: string; provider: string; accessToken: string },
+  ) {
+    try {
+      const linkedUser = await this.userService.linkSocialAccount(userId, provider, accessToken);
+      return linkedUser;
+    } catch (error) {
+      if (error instanceof HttpException) {
+        throw error; // Re-throw existing HttpExceptions for consistent error handling
+      } else {
+        throw new HttpException('Failed to link social account', HttpStatus.INTERNAL_SERVER_ERROR);
+      }
+    }
   }
 }
