+++ backend/src/user/service/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { User } from '@prisma/client';
 import { PrismaService } from '../../prisma/prisma.service';
 import { RatingDto } from '../dto/rating.dto';
@@ -75,5 +75,42 @@ export class UserService {
       createdOn: review.createdAt.toISOString(),
       comment: review.comment,
     }));
+  }
+
+  async linkSocialAccount(userId: string, provider: string, accessToken: string) {
+    const user = await this.findUserById(userId);
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    let email: string | undefined;
+
+    const existingUser = await this.findUserByEmail(email);
+    if (existingUser && existingUser.id !== user.id) {
+      throw new HttpException('Social account already linked to a different user', HttpStatus.CONFLICT);
+    }
+
+    user.socialAccounts = user.socialAccounts || [];
+    user.socialAccounts.push({ provider, accessToken });
+
+    await this.updateUser(user);
+
+    return user;
+  }
+
+  private async findUserById(id: string) {
+    return await this.prisma.user.findUnique({ where: { id } });
+  }
+
+  private async updateUser(user: any) {
+    return await this.prisma.user.update({ where: { id: user.id }, data: user });
+  }
+
+  private async findUserByEmail(email: string) {
+    return await this.prisma.user.findUnique({
+      where: {
+        email,
+      },
+    });
   }
 }
